fix bug for deleting potential non-exist folder; add check of set up path for generic single job submission function

+++ microscopeDataProcessing/stitch/XR_stitching_frame_zarr_dev_v1.m
@@ -891,7 +891,7 @@ if ~exist(nv_fullname, 'dir') && ~all(is_done_flag)
     end
 end
 
-if ~debug
+if ~debug && exist(zarrFlagPath, 'dir')
     rmdir(zarrFlagPath, 's');
 end
